requirements.txt
essentia==2.1b6.dev1032
mysql-connector-python==8.0.32
python-telegram-bot==13.11